auth-local/requirements.txt
Django~=5.1.6
psycopg2~=2.9.10
djangorestframework
django-cors-headers==3.11.0
requests
PyJWT==2.8.0
django-environ==0.4.5
psycopg2-binary
python-dotenv
django-prometheus
gunicorn
django-otp==1.2.2
qrcode==7.4.2
Pillow==10.0.0